# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: IITM Beacon API
  description: >
    Exchange Student Testimonial & Review Gallery. Covers the six feature slices: gallery
    (public browse/filter/search), submission (visitor OTP login, session ping,
    create/edit-testimonial), moderation (admin queue/approve/reject, one at a time, session
    ping), catalogadmin (admin CRUD for topic groups/topics/achievements), adminauth
    (passwordless OTP login), analytics (homepage dashboard). Revealing a testimonial's public
    contact methods (UC-REVEAL-CONTACT) is deliberately not part of this JSON API: it is only
    served to the site's own article page, as the HTML route `POST /gallery/{id}/contact`
    (decision 27). See ../use-cases.md for the scenarios behind each endpoint,
    ../decisions.md for the design decisions behind each schema shape, and ../nfr.md for the
    quality attributes they must meet.
  version: "0.3.0"
servers:
  - url: /api
...
paths:
  /gallery/testimonials:
    get:
      tags: [gallery]
      summary: List approved testimonials (UC-BROWSE-APPROVED, UC-FILTER-COUNTRY, UC-SEARCH-KEYWORD, UC-FILTER-TOPIC)
      parameters:
        - name: country
          in: query
          description: Filter by country code (ISO 3166-1 alpha-2, decision 1).
          schema: { type: string, minLength: 2, maxLength: 2 }
        - name: topicIds
          in: query
          description: >
            Filter by one or more top-level topic-catalog entries — a topic-group id or a
            standalone topic id (decision 11, UC-FILTER-TOPIC). A group id expands
            server-side to all of its member topic ids.
          schema:
            type: array
            items: { type: integer }
        - name: q
          in: query
          description: >
            Free-text search over section answer text and the submitter's first/last name
            (decision 8, decision 10).
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Paginated list of approved testimonials.
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/TestimonialCard" }
                  page: { type: integer }
                  size: { type: integer }
                  totalElements: { type: integer }
  /gallery/testimonials/{id}:
    get:
      tags: [gallery]
      summary: Get full testimonial detail (UC-EXPAND-TESTIMONIAL)
      description: >
        Never includes contact-method values or the submitter's email (decision 6) — only
        `hasRevealableContact`, telling whether the testimonial has any public contact method
        (the values themselves are revealed only on the article page — decision 27).
        `displayName` (decision 10) appears only here, never on `TestimonialCard`.
      parameters:
        - $ref: "#/components/parameters/TestimonialId"
      responses:
        "200":
          description: Testimonial detail, rendered as an article of its filled sections.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TestimonialDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /submissions:
    post:
      tags: [submission]
      summary: Create a new testimonial (UC-CREATE-TESTIMONIAL)
      security: [{ visitorSession: [] }]
      description: >
        Topics are database-driven (decision 11), so the request cannot have one named field
        per topic. `payload` is a JSON part; photo files are separate multipart parts,
        cross-referenced from `payload.sections[].photos[].fileRef` by index. The exact
        multipart framing is an implementation detail; this schema fixes the logical shape
        only.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/TestimonialSubmission" }
      responses:
        "201":
          description: Saved with status PENDING.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  status: { type: string, enum: [PENDING] }
        "400":
          description: >
            Validation failure — no section filled in, missing recommendation score, a
            mandatory identity field missing, no country ("Please select your country."), a
            roll number not in the `AA00A000` shape or an admission year outside
            1959..current year (decision 10), data-processing consent unchecked (decision 7),
            photos on a section with a blank answer, a contact value not matching its type's
            `valuePattern` (decision 5), more photos on one section than the per-topic limit
            ("At most 5 photos per topic.") or on the testimonial than the total limit ("Too
            many photos: maximum is 50."; defaults, decision 22), more than 10 custom tags on
            one photo, or a multipart request the server can't parse (e.g. more parts than the
            configured part-count limit). Every one of these rules the request breaks is
            reported at once, as `field: message` pairs joined into `message`. Only once they
            all pass are the photo files themselves checked and converted, one at a time, so
            the first photo that fails is reported alone: over the size limit ("Photo exceeds
            the maximum allowed file size (20 MB)."), not a readable image ("Unsupported image
            format. Please upload JPEG, PNG, WebP, GIF, TIFF or BMP." — HEIC/HEIF and AVIF
            included), or above the pixel limit ("The image is too large: at most 250
            megapixels are allowed.").
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: A testimonial already exists for this visitor — use PUT /submissions/mine instead.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "413":
          description: >
            The multipart request exceeds the server's upload limits (per-part or per-request
            size; defaults 25MB and 1010MB, above the app's own 20 MB photo limit) before the
            application's own validation can run.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /submissions/mine:
    get:
      tags: [submission]
      summary: Load the authenticated visitor's own testimonial for editing (UC-EDIT-TESTIMONIAL)
      security: [{ visitorSession: [] }]
      responses:
        "200":
          description: The visitor's existing testimonial, pre-filled into the same shape used to submit it.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TestimonialSubmission" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No testimonial exists yet for this visitor — use POST /submissions instead.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    put:
      tags: [submission]
      summary: Edit the authenticated visitor's existing testimonial (UC-EDIT-TESTIMONIAL)
      security: [{ visitorSession: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/TestimonialSubmission" }
      responses:
        "200":
          description: >
            `status` reflects whether the moderation short-circuit applied (decision 18):
            `APPROVED` means the edit changed only exempt fields (country, score, achievements,
            contact methods) on a testimonial that was already `APPROVED`, and is live
            immediately; `PENDING` means it re-entered moderation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  status: { type: string, enum: [PENDING, APPROVED] }
        "400":
          description: Same validation failures as POST /submissions.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No testimonial exists yet for this visitor — use POST /submissions instead.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "413":
          description: >
            The multipart request exceeds the server's upload limits (per-part or per-request
            size; defaults 25MB and 1010MB, above the app's own 20 MB photo limit) before the
            application's own validation can run.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /moderation/testimonials/pending:
    get:
      tags: [moderation]
      summary: List pending testimonials (UC-VIEW-PENDING-QUEUE)
      security: [{ adminSession: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: >
            Paginated list of PENDING testimonials, full detail including decrypted contact
            methods and email, plus the admin-only diff flags (decision 18).
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/ModerationTestimonialDetail" }
                  page: { type: integer }
                  size: { type: integer }
                  totalElements: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /moderation/testimonials/{id}/approve:
    post:
      tags: [moderation]
      summary: Approve one pending testimonial (UC-APPROVE-TESTIMONIAL)
      security: [{ adminSession: [] }]
      parameters:
        - $ref: "#/components/parameters/TestimonialId"
      responses:
        "204": { description: Status set to APPROVED; every modified flag cleared. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /moderation/testimonials/{id}/reject:
    post:
      tags: [moderation]
      summary: Reject one pending testimonial (UC-REJECT-TESTIMONIAL)
      security: [{ adminSession: [] }]
      parameters:
        - $ref: "#/components/parameters/TestimonialId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, nullable: true }
      responses:
        "204":
          description: >
            Status set to REJECTED, `rejectedAt` stamped. The submitter is always emailed to
            let them know to log in and fix it; the reason is included in that email when one
            was given.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /moderation/session:
    get:
      tags: [moderation]
      summary: Session ping for the admin's pages (decision 23)
      security: [{ adminSession: [] }]
      description: >
        Answers whether the admin session is still alive; the request itself extends it.
        Called by the admin pages' session check (static/js/session-check.js) whenever their
        tab becomes visible again — on 401 or 403 the page reloads and so goes through the
        login redirect. Responses are not cacheable (`Cache-Control` includes `no-store`).
      responses:
        "204": { description: The admin session is alive. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Established via /admin/auth/otp/verify; principal holds the admin role.
    visitorSession:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: >
        Established via /submissions/otp/verify; principal holds the visitor role, scoped to
        the one email that was verified (decision 17).
  parameters:
    TestimonialId:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: No valid session for the required role.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: A valid session, but of the other role (e.g. an admin session on a visitor endpoint).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    Country:
      type: object
      properties:
        code: { type: string, minLength: 2, maxLength: 2, description: "ISO 3166-1 alpha-2" }
        name: { type: string }
...
    PhotoRef:
      description: >
        One photo of a section. Uploads are stored as a full-size WebP plus a thumbnail WebP
        (decision 22). A legacy photo stored before that pipeline, until converted, has its
        original file as both `url` and `thumbnailUrl` and null `width`/`height`.
      type: object
      properties:
        url: { type: string, description: "The full-size image." }
        thumbnailUrl: { type: string, description: "The thumbnail (at most 640 px on the long edge by default)." }
        width:
          type: integer
          nullable: true
          description: Full-size image width in pixels; null for a legacy photo.
        height:
          type: integer
          nullable: true
          description: Full-size image height in pixels; null for a legacy photo.
        tags:
          type: array
          items: { type: string }
          maxItems: 10
          description: Free-form custom tags, in addition to the section's topic tag (decision 2).
    TestimonialSectionView:
      type: object
      properties:
        topicSlug: { type: string }
        topicLabel: { type: string }
        answer: { type: string }
        photos:
          type: array
          items: { $ref: "#/components/schemas/PhotoRef" }
        updated:
          type: boolean
          description: >
            True if this section's text or photos changed since the last approval
            (`TestimonialSection.modified`, decision 18) — drives the public "updated" badge
            (UC-EXPAND-TESTIMONIAL) and, for the admin, the moderation-queue diff.
    TestimonialCard:
      description: >
        Shape returned by the gallery list endpoint (UC-BROWSE-APPROVED). Deliberately carries
        no name-derived field at all (decision 10).
      type: object
      properties:
        id: { type: integer }
        country: { $ref: "#/components/schemas/Country" }
        recommendationScore: { type: integer, minimum: 0, maximum: 10 }
        previewText:
          type: string
          description: Drawn from the first filled section, for the card preview.
        thumbnailUrl:
          type: string
          nullable: true
          description: >
            The card cover — the thumbnail of the first photo of the first filled section (its
            original file for a legacy photo); null when that section has no photo.
    TestimonialDetail:
      description: >
        Full public testimonial detail, rendered as an article of its filled sections. Never
        includes contact-method values or the email (decision 6) — only
        `hasRevealableContact`. `displayName` appears only here, never on `TestimonialCard`
        (decision 10).
      type: object
      properties:
        id: { type: integer }
        displayName:
          type: string
          description: "First name + last-initial, e.g. \"David J.\" (decision 10)."
        country: { $ref: "#/components/schemas/Country" }
        recommendationScore: { type: integer, minimum: 0, maximum: 10 }
        status: { type: string, enum: [PENDING, APPROVED, REJECTED] }
        sections:
          type: array
          items: { $ref: "#/components/schemas/TestimonialSectionView" }
        achievements:
          type: array
          items: { type: string }
          description: Achievement slugs the submitter checked.
        hasRevealableContact:
          type: boolean
          description: >
            Whether the testimonial has at least one contact method marked public — i.e. whether
            its article page offers "Reveal contact info" (UC-REVEAL-CONTACT, decision 27).
    ModerationContactMethodView:
      description: Shape used in moderation views — includes every entry, public or private.
      type: object
      properties:
        type: { type: string }
        value: { type: string }
        isPublic: { type: boolean }
    ModerationTestimonialDetail:
      description: >
        Full detail as seen by the admin (UC-VIEW-PENDING-QUEUE) — includes identity fields,
        the decrypted email, every contact method (decisions 6, 10), and the admin-only diff
        flags (decision 18).
      type: object
      properties:
        id: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        rollNumber: { type: string }
        admissionYear: { type: integer }
        email: { type: string }
        country: { $ref: "#/components/schemas/Country" }
        recommendationScore: { type: integer, minimum: 0, maximum: 10 }
        status: { type: string, enum: [PENDING, APPROVED, REJECTED] }
        sections:
          type: array
          items: { $ref: "#/components/schemas/TestimonialSectionView" }
        achievements:
          type: array
          items: { type: string }
        contactMethods:
          type: array
          items: { $ref: "#/components/schemas/ModerationContactMethodView" }
        identityModified:
          type: boolean
          description: True if any identity field changed since the last approval — always forced this item into the queue (decision 18).
        scoreModified:
          type: boolean
          description: True if recommendationScore changed since the last approval — shown as context only, never why this item is here (decision 18).
    TestimonialSubmission:
      description: >
        Logical shape of the multipart submission body for POST /submissions and
        PUT /submissions/mine, and the response body of GET /submissions/mine.
      type: object
      required:
        [firstName, lastName, rollNumber, admissionYear, countryCode, recommendationScore,
         sections, dataProcessingConsent]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        rollNumber:
          type: string
          pattern: "^[A-Za-z]{2}[0-9]{2}[A-Za-z][0-9]{3}$"
          description: >
            IITM roll number in the `AA00A000` shape, e.g. `CS21B001` (decision 10). Trimmed and
            uppercased before validation, so lowercase letters are accepted and stored uppercased.
        admissionYear:
          type: integer
          minimum: 1959
          description: Between 1959 and the current year inclusive (decision 10).
        countryCode: { type: string, minLength: 2, maxLength: 2 }
        recommendationScore: { type: integer, minimum: 0, maximum: 10 }
        sections:
          type: array
          minItems: 1
          description: >
            At least one filled section required (decision 12). Each references a live Topic by
            slug. A section whose `answer` is blank doesn't count as filled and is dropped, but
            only if it carries no photos — photos on a blank section are rejected with 400.
          items:
            type: object
            required: [topicSlug, answer]
            properties:
              topicSlug: { type: string }
              answer: { type: string }
              photos:
                type: array
                description: >
                  Indexes into the multipart file parts carrying this section's photos. At most
                  the configured per-topic limit per section and the total limit per testimonial,
                  kept and new photos counted together (defaults 5 and 50, decision 22). Each new
                  file is converted to WebP on save; the uploaded file itself is never stored.
                items:
                  type: object
                  properties:
                    fileRef:
                      type: string
                      description: >
                        On create, always names a multipart file part (a new upload). On edit
                        (PUT /submissions/mine), it is dual-purpose: if it matches an existing
                        photo's own `url` (as returned by GET /submissions/mine, in this same
                        section) that photo is kept unchanged except for `tags`; otherwise it
                        must name a multipart file part, i.e. a genuinely new upload. Any of the
                        section's existing photos whose `url` isn't referenced this way are
                        removed (decision 18).
                    tags:
                      type: array
                      items: { type: string }
                      maxItems: 10
        achievementSlugs:
          type: array
          items: { type: string }
          description: Achievement checkboxes ticked, from the current active Achievement list.
        contactMethods:
          type: array
          description: 0-N contact methods (decision 5), each independently marked public/private.
          items:
            type: object
            required: [typeSlug, value, isPublic]
            properties:
              typeSlug: { type: string, description: "ContactType slug, e.g. \"telegram\"" }
              value:
                type: string
                description: >
                  Trimmed, then must fully match the type's `valuePattern` when it has one
                  (decision 5); otherwise only non-blank.
              isPublic: { type: boolean, default: false }
        dataProcessingConsent:
          type: boolean
          description: Must be true, per decision 7 — request is rejected otherwise.
...
    ErrorResponse:
      description: Uniform error shape returned by @RestControllerAdvice (per CLAUDE.md baseline).
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        path: { type: string }